refactor: move 27 manifest toolchain-config types to homeboy-extension-contract
Eighth sub-cut of the extension-contract campaign (Stage 1, wiki 12840) — the
largest yet, and it unblocks the DeployCapability sub-tree.

Analysis showed manifest_config.rs (29 types) was blocked from moving only by
LifecycleContract (which chains through core artifact_contract -> observation).
But LifecycleContract is carried by exactly ONE type, FuzzWorkloadConfig (and
transitively FuzzConfig). The other 27 types are lifecycle-free.

Moves those 27 into a new contract manifest_toolchain_config module (deploy,
build, test, lint, cli, database, discovery, requirements, source-snapshot
config types + the default_staging_path serde helper), leaving FuzzConfig and
FuzzWorkloadConfig in core. This also unblocks manifest.rs's DeployCapability,
whose fields (DeployVerification, DeployOverride, RemotePathInferenceRule,
VersionPatternConfig, SinceTagConfig, etc.) now resolve from the contract crate.

Added homeboy-engine-primitives as a direct contract dep (for ParseSpec).
manifest.rs re-exports the moved types, keeping crate::extension::* stable.
