+++ crates/homeboy-extension-contract/src/lib.rs
@@ -17,6 +17,7 @@ pub mod manifest_action_config;
 pub mod manifest_capability_config;
 pub mod manifest_deploy_config;
 pub mod manifest_test_config;
+pub mod manifest_toolchain_config;
 pub mod runner_contract;
 pub mod source_metadata_repair;
 pub mod test_drift;
